// src/app/models/product.ts
export default class Products {
    id_product: string;
    name: string;
    category: number;
    first_price: number;
    last_price: number;
    step_price: number;
    day_start: string;
    day_end: string;
    image: string;
    status: number;
    description: string;
    id_seller:number;
    id_bidder:number;
    current_price:number;
}